fortinet-api/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
sqlalchemy>=2.0.0
pydantic>=2.0.0
psycopg2-binary>=2.9.5
python-dotenv>=1.0.0
